+++ module/utility/chat-dice.mjs
@@ -1,13 +1,89 @@
 import { HeroRoller } from "./dice.mjs";
 import { HEROSYS } from "../herosystem6e.mjs";
 
-const heroRollRegExp = new RegExp(
-    "^(?<cmd>\\/heroroll)(?:[\\s]+)(?<nonCmd>(?<numDice>[\\d\\.]+)d(?<diceSize>[\\d]+)?(?<numTerm>(?<numTermSign>[-+]?)[\\d]+)?(?<flavourTerm>\\[(?<flavourTermContent>(?<heroSystemVersion>[56]?)(?<hitLoc>h)?(?<flavour>.*))\\])?)$",
-    "i",
-);
+const heroRollRegExpString =
+    "(?<cmd>\\/heroroll)(?:[\\s]+)(?<nonCmd>(?<numDice>[\\d\\.]+)d(?<diceSize>[\\d]+)?(?<numTerm>(?<numTermSign>[-+]?)[\\d]+)?(?<flavourTerm>\\[(?<flavourTermContent>(?<heroSystemVersion>[56]?)(?<hitLoc>h)?(?<flavour>.*))\\])?)";
+const chatHeroRollRegExpString = `^${heroRollRegExpString}$`;
+const inlineHeroRollRegExpStr = `\\[\\[${heroRollRegExpString}\\]\\]`;
 
+const chatHeroRollRegExp = new RegExp(chatHeroRollRegExpString, "i");
+const inlineHeroRollRegExp = new RegExp(inlineHeroRollRegExpStr, "gi");
+
+async function onInlineHeroRollClick(event) {
+    const a = event.target.closest("a.inline-hero-roll");
+    if (!a) {
+        return;
+    }
+    event.preventDefault();
+
+    // Recover the data stashed in data-* attributes
+    const match = JSON.parse(a.dataset.match);
+    match.groups = JSON.parse(a.dataset.groups);
+    match.input = a.dataset.input;
+
+    const heroRoller = buildHeroRoller(match);
+    await rollAndGenerateChatMessage(heroRoller);
+}
+
+/**
+ * Text Editor Enricher (Of type TextEditorEnricher) to provide inline support for /heroroll
+ *
+ * @param {RegExpMatchArray} match
+ * @param {EnrichmentOptions} [options]
+ *
+ * @returns [Promise<HTMLElement | null>]
+ */
+async function heroRollTextEditorEnricher(match /*, options*/) {
+    if (match && match[0]) {
+        // Build a clickable anchor element
+        const a = document.createElement("a");
+        a.classList.add("inline-hero-roll", "roll");
+        a.dataset.match = JSON.stringify(match);
+        a.dataset.groups = JSON.stringify(match.groups);
+        a.dataset.input = JSON.stringify(match.input);
+
+        const heroRoller = buildHeroRoller(match);
+
+        // Add a link that looks similar to Foundry's inline roll link
+        const linkActionDescription = `${match[2].replace(/\[.*\]/, "")} ${heroRoller.getType()} roll ${heroRoller.hitLocationValid() ? " with hit location" : ""}`;
+        a.innerHTML = `<i class="fas fa-dice-d6"></i> ${linkActionDescription}`;
+
+        // Evaluate roll when link is clicked. This will be stripped out in chat messages but presumably won't be elsewhere this might be used - but I'm actually not sure.
+        a.addEventListener("click", onInlineHeroRollClick);
+
+        return a;
+    }
+
+    return null;
+}
+
+/**
+ * Register an inline hero roller text editor enricher.
+ */
+Hooks.on("init", function () {
+    CONFIG.TextEditor.enrichers.push({
+        enricher: heroRollTextEditorEnricher,
+        pattern: inlineHeroRollRegExp,
+
+        // Don't seem to need these, but they are there as options
+        // id?: string;
+        // onRender?: (arg0: HTMLEnrichedContentElement) => any;
+        // replaceParent?: boolean;
+    });
+});
+
+/**
+ * When chat messages are rendered, see if there are any inline /heroroll rolls. If so, register a click handler for it.
+ */
+Hooks.on("renderChatMessage", (message, html) => {
+    html[0].querySelectorAll("a.inline-hero-roll").forEach((a) => a.addEventListener("click", onInlineHeroRollClick));
+});
+
+/**
+ * When a chat message is generated, check if this is a /heroroll command.
+ */
 Hooks.on("chatMessage", function (_this, message /*, _chatData*/) {
-    const chatMessageCmd = message.match(heroRollRegExp);
+    const chatMessageCmd = message.match(chatHeroRollRegExp);
     if (chatMessageCmd?.groups?.cmd.toLowerCase() === "/heroroll") {
         // Minimal error handling: Confirm "d6" or it's just a straight numerical term
         if (
@@ -29,10 +105,14 @@ Hooks.on("chatMessage", function (_this, message /*, _chatData*/) {
     return true;
 });
 
-async function doRollAndGenerateChatMessage(chatMessageCmd) {
-    const useHitLocations = !!chatMessageCmd.groups.hitLoc;
-    const useHitLocationsSide = game.settings.get(HEROSYS.module, "hitLocTracking") === "all";
-
+/**
+ * Generate a HeroRoller object from the /heroroll command
+ *
+ * @param {RegExpMatch} chatMessageCmd
+ *
+ * @returns {HeroRoller}
+ */
+function buildHeroRoller(chatMessageCmd) {
     let numericTerm = parseFloat(chatMessageCmd.groups.numTerm || 0);
     const negativeTermWithDice = chatMessageCmd.groups.numDice && numericTerm < 0; // e.g. 1d6-1
     if (negativeTermWithDice) {
@@ -47,11 +127,8 @@ async function doRollAndGenerateChatMessage(chatMessageCmd) {
             : 0,
     );
 
-    const roller = new HeroRoller()
-        .modifyTo5e(chatMessageCmd.groups.heroSystemVersion === "5")
-        .addToHitLocation(useHitLocations, "none", useHitLocationsSide, "none");
+    const heroRoller = new HeroRoller().modifyTo5e(chatMessageCmd.groups.heroSystemVersion === "5");
 
-    let flavour;
     switch (chatMessageCmd.groups.flavour?.toLowerCase()) {
         case "k":
             {
@@ -60,8 +137,7 @@ async function doRollAndGenerateChatMessage(chatMessageCmd) {
                     "NonStandardStunMultiplierForKillingAttackBackingSetting",
                 );
 
-                flavour = HeroRoller.ROLL_TYPE.KILLING;
-                roller.makeKillingRoll(
+                heroRoller.makeKillingRoll(
                     true,
                     customStunMultiplierSetting.d6Count ||
                         customStunMultiplierSetting.d6Less1DieCount ||
@@ -74,60 +150,83 @@ async function doRollAndGenerateChatMessage(chatMessageCmd) {
             break;
 
         case "n":
-            flavour = HeroRoller.ROLL_TYPE.NORMAL;
-            roller.makeNormalRoll();
+            heroRoller.makeNormalRoll();
             break;
 
         case "a":
-            flavour = HeroRoller.ROLL_TYPE.ADJUSTMENT;
-            roller.makeAdjustmentRoll();
+            heroRoller.makeAdjustmentRoll();
             break;
 
         case "e":
-            flavour = HeroRoller.ROLL_TYPE.ENTANGLE;
-            roller.makeEntangleRoll();
+            heroRoller.makeEntangleRoll();
             break;
 
         case "s":
-            flavour = HeroRoller.ROLL_TYPE.SUCCESS;
-            roller.makeSuccessRoll();
+            heroRoller.makeSuccessRoll();
             break;
 
         case "f":
-            flavour = HeroRoller.ROLL_TYPE.FLASH;
-            roller.makeFlashRoll();
+            heroRoller.makeFlashRoll();
+            break;
+
+        case "l":
+            heroRoller.makeLuckRoll();
+            break;
+
+        case "u":
+            heroRoller.makeUnluckRoll();
             break;
 
         default:
-            flavour = HeroRoller.ROLL_TYPE.BASIC;
-            roller.makeBasicRoll();
+            heroRoller.makeBasicRoll();
             break;
     }
 
-    // Capitalize the first letter
-    flavour = `${flavour.charAt(0).toUpperCase() + flavour.slice(1)} attack`;
-
-    roller
+    heroRoller
         .addDice(Math.trunc(numDice))
         .addDiceMinus1(negativeTermWithDice ? 1 : 0)
         .addHalfDice(numDice % 1 ? 1 : 0)
         .addNumber(Math.trunc(numericTerm));
 
-    await roller.roll();
+    const useHitLocations = !!chatMessageCmd.groups.hitLoc;
+    const useHitLocationsSide = game.settings.get(HEROSYS.module, "hitLocTracking") === "all";
+    heroRoller.addToHitLocation(useHitLocations, "none", useHitLocationsSide, "none");
 
-    if (useHitLocations) {
-        flavour += ` to ${roller.getHitLocation().fullName}`;
-    }
-    const cardHtml = await roller.render(flavour);
+    return heroRoller;
+}
+
+/**
+ * Evaluate the hero roller and generate a chat card for it.
+ *
+ * @param {HeroRoller} heroRoller
+ */
+async function rollAndGenerateChatMessage(heroRoller) {
+    await heroRoller.roll();
+
+    // Setup flavour text with capitalized first letter
+    const chatCardFlavour = `${heroRoller.getType().charAt(0).toUpperCase() + heroRoller.getType().slice(1)} attack ${heroRoller.hitLocationValid() ? ` to ${heroRoller.getHitLocation().fullName}` : ""}`;
+    const cardHtml = await heroRoller.render(chatCardFlavour);
 
     const speaker = ChatMessage.getSpeaker();
     const chatData = {
         style: CONST.CHAT_MESSAGE_STYLES.OOC,
-        rolls: roller.rawRolls(),
+        rolls: heroRoller.rawRolls(),
         author: game.user._id,
         content: cardHtml,
         speaker: speaker,
     };
 
     await ChatMessage.create(chatData);
+}
+
+/**
+ * Parse the /heroroll command and evaluate it generating a chat card with the result.
+ *
+ * @param {RegExpMatch} chatMessageCmd
+ *
+ * @returns
+ */
+async function doRollAndGenerateChatMessage(chatMessageCmd) {
+    const heroRoller = buildHeroRoller(chatMessageCmd);
+    return rollAndGenerateChatMessage(heroRoller);
 }